third_party/WebKit/LayoutTests/media/encrypted-media/encrypted-media-waiting-for-a-key-expected.txt

EVENT(encrypted)
EVENT(encrypted)
EVENT(message)
EVENT(playing)
This is a testharness.js-based test.
PASS Waiting for a key. 
Harness: the test ran to completion.
